Add AssertBlock to marked lines
