+++ flake8_aaa/assert_block.py
@@ -1,11 +1,12 @@
-class AssertBlock(object):
-    """
-    Attributes:
-        nodes (list (ast Node))
-    """
+import ast
 
-    def __init__(self):
-        self.nodes = []
+from .multi_node_block import MultiNodeBlock
+from .types import LineType
 
-    def add_node(self, node):
+
+class AssertBlock(MultiNodeBlock):
+    line_type = LineType.assert_block
+
+    def add_node(self, node: ast.AST) -> bool:
         self.nodes.append(node)
+        return True
